requirements.txt
peewee
slackclient
chalice
PyJWT